requirements.txt
requests
fastapi
uvicorn[standard]
python-telegram-bot[httpx]
pyjwt